Refactored streaming functionality of StreamActivity into a fragment and fix camera issues such as rotation and release
