+++ client/android/WorldScope/app/src/main/res/layout/activity_stream.xml
@@ -1,11 +1,24 @@
 <?xml version="1.0" encoding="utf-8"?>
 <FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools" android:layout_width="match_parent"
-    android:id="@+id/streamFrameLayout"
-    android:layout_height="match_parent" android:paddingLeft="@dimen/activity_horizontal_margin"
-    android:paddingRight="@dimen/activity_horizontal_margin"
-    android:paddingTop="@dimen/activity_vertical_margin"
-    android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context="com.litmus.worldscope.StreamActivity">
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_height="fill_parent"
+    android:layout_width="fill_parent"
+    tools:context="com.litmus.worldscope.StreamActivity"
+    android:background="#ff0000">
+
+    <fragment
+        android:layout_width="fill_parent"
+        android:layout_height="249dp"
+        android:name="layout.StreamVideoFragment"
+        android:id="@+id/streamVideoFragment"
+        android:layout_gravity="center_horizontal|bottom" />
+
+    <TextView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:textAppearance="?android:attr/textAppearanceLarge"
+        android:text="HELLO"
+        android:id="@+id/textView"
+        android:layout_gravity="center" />
 
 </FrameLayout>
