+++ client/android/WorldScope/app/src/main/res/layout/fragment_stream_video.xml
@@ -1,0 +1,7 @@
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/streamFrameLayout"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context="layout.StreamVideoFragment">
+</FrameLayout>
